+++ src/core/tools/WriteToFileTool.ts
@@ -282,7 +282,7 @@ export class WriteToFileTool extends BaseTool<"write_to_file"> {
 		const relPath: string | undefined = block.params.path
 		let newContent: string | undefined = block.params.content
 
-		if (!relPath || newContent === undefined) {
+		if (!relPath) {
 			return
 		}
 
@@ -313,7 +313,7 @@ export class WriteToFileTool extends BaseTool<"write_to_file"> {
 		const sharedMessageProps: ClineSayTool = {
 			tool: fileExists ? "editedExistingFile" : "newFileCreated",
 			path: getReadablePath(task.cwd, relPath),
-			content: newContent,
+			content: newContent || "",
 			isOutsideWorkspace,
 			isProtected: isWriteProtected,
 		}
@@ -321,14 +321,16 @@ export class WriteToFileTool extends BaseTool<"write_to_file"> {
 		const partialMessage = JSON.stringify(sharedMessageProps)
 		await task.ask("tool", partialMessage, block.partial).catch(() => {})
 
-		if (!task.diffViewProvider.isEditing) {
-			await task.diffViewProvider.open(relPath)
-		}
+		if (newContent) {
+			if (!task.diffViewProvider.isEditing) {
+				await task.diffViewProvider.open(relPath)
+			}
 
-		await task.diffViewProvider.update(
-			everyLineHasLineNumbers(newContent) ? stripLineNumbers(newContent) : newContent,
-			false,
-		)
+			await task.diffViewProvider.update(
+				everyLineHasLineNumbers(newContent) ? stripLineNumbers(newContent) : newContent,
+				false,
+			)
+		}
 	}
 }